add comment list

+++ src/main/java/life/sanelee/community/Controller/PublishController.java
@@ -1,7 +1,6 @@
 package life.sanelee.community.Controller;
 
 import life.sanelee.community.dto.QuestionDTO;
-import life.sanelee.community.mapper.QuestionMapper;
 import life.sanelee.community.model.Question;
 import life.sanelee.community.model.User;
 import life.sanelee.community.service.QuestionService;
@@ -22,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
+    public String edit(@PathVariable(name="id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -43,7 +42,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
